requirements.txt
python-dotenv
pandas
numpy
setuptools
pymongo[srv]==3.12
certifi
scikit-learn
mlflow==2.16.0
pyaml
dagshub
fastapi
uvicorn
jinja2
python-multipart
aiofiles

# -e .